Update calculator.inc.php
I have modified the code sent to separate the original code and the code created for the rectangle tank calculator.  This will hopefully help to prevent any confusion as to what functions are being called and why.  The new code adds a main if statement to the original code block completely separating it from the new code block.  This is a movement toward procedural programing, however; and is not optimal.